requirements.txt
openai
azure.identity
azure.ai.projects
azure.ai.agents